Added slow-spin class to allow icon to be set to a slower spin speed. Added links to my work on codepen using codepins twitter image meta tags
Changed html to allow for spinning icon.
Changed icon to be sub-spanned to allow for a box to be rotated (to give cog spinning effect around the circle).
Changed Main header content (header and byline) to be more personalize.
Changed intro sub menu to have a picture of me.
Changed display type in intro sub menu to flex and made image inwrap (using float) with the text that is set.
Changed Work menu to be a gallery format (adding gallery container and item classes).