Update index.html
When on title page with small width (aspect ratio 0.5<) it could make Experience and optimization appear on the same line. When this happen they appeared as a combined word experinceoptimization. The code is experience<br>optimization but because br only respects if there is enough space to validate this it would create and an issue where these words appeared together.
Updated by adding a space between experience and <br>.